PROJECT-1/requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
google-generativeai
pypdf
faiss-cpu
python-dotenv
